Remove Duplicate Code On UnitTests
