+++ Test/BSN.Commons.Orm.EntityFramework.Tests/Mock/UsersRepsitory.cs
@@ -1,6 +1,7 @@
 ﻿using BSN.Commons.Test.Data;
 using BSN.Commons.Infrastructure;
 using BSN.Commons.Orm.EntityFramework;
+using BSN.Commons.Tests;
 
 namespace BSN.Commons.Test.Mock
 {
