+++ Test/BSN.Commons.Orm.EntityFrameworkCore.Tests/BSN.Commons.Orm.EntityFrameworkCore.Tests.csproj
@@ -28,6 +28,7 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Source\BSN.Commons.Orm.EntityFrameworkCore\BSN.Commons.Orm.EntityFrameworkCore.csproj" />
     <ProjectReference Include="..\..\Source\BSN.Commons\BSN.Commons.csproj" />
+    <ProjectReference Include="..\BSN.Commons.Tests\BSN.Commons.Tests.csproj" />
   </ItemGroup>
 
 </Project>
