--- Test/BSN.Commons.Orm.EntityFrameworkCore.Tests/Data/User.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace BSN.Commons.Test.Data
-{
-    public class User
-    {
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public long Id { get; set; }
-
-        [MaxLength(7)]
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public Document Document { get; set; }
-        public string Password { get; set; }
-    }
-}
