+++ Test/BSN.Commons.Orm.EntityFrameworkCore.Tests/Mock/UsersRepsitory.cs
@@ -1,6 +1,7 @@
 ﻿using BSN.Commons.Orm.EntityFrameworkCore;
 using BSN.Commons.Test.Data;
 using BSN.Commons.Infrastructure;
+using BSN.Commons.Tests;
 
 namespace BSN.Commons.Test.Mock
 {
