+++ Test/BSN.Commons.Tests/BSN.Commons.Tests.csproj
@@ -22,4 +22,12 @@
     <ProjectReference Include="..\..\Source\BSN.Commons\BSN.Commons.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="obj\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Reference Include="System.Transactions" />
+  </ItemGroup>
+
 </Project>
