--- Test/BSN.Commons.Tests/Class1.cs
@@ -1,18 +1,0 @@
-﻿using NUnit.Framework;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BSN.Commons.Tests
-{
-    class Class1
-    {
-        [Test]
-        public void Test()
-        {
-            Assert.True(true);
-        }
-    }
-}
