+++ Test/BSN.Commons.Tests/UnitOfWorkArrangementsAbstractFactory.cs
@@ -1,0 +1,11 @@
+﻿using BSN.Commons.Infrastructure;
+
+namespace BSN.Commons.Tests
+{
+    public abstract class UnitOfWorkArrangementsAbstractFactory
+    {
+        public abstract IDatabaseFactory CreateDatabaseFactory();
+        public abstract IRepository<User> CreateUserRepository(IDatabaseFactory databaseFactory);
+    }
+
+}
