+++ Test/BSN.Commons.Tests/User.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace BSN.Commons.Test.Data
+namespace BSN.Commons.Tests
 {
     public class User
     {
